elf-loader: compile-time api hash collision validation
