+++ applications/elf-loader/elf-cpp/elf-hashtable.cpp
@@ -1,5 +1,7 @@
 #include "compilesort.hpp"
 #include "elf-hashtable.h"
+#include "elf-hashtable-entry.h"
+#include "elf-hashtable-checks.hpp"
 
 #include <gui/gui.h>
 #include <furi.h>
@@ -8,26 +10,6 @@
 #include <algorithm>
 
 #define TAG "elf-hashtable"
-
-#define API_METHOD(x, ret_type, args_type)                                                     \
-    sym_entry {                                                                                \
-        .hash = elf_gnu_hash(#x), .address = (uint32_t)(static_cast<ret_type(*) args_type>(x)) \
-    }
-
-struct sym_entry {
-    uint32_t hash;
-    uint32_t address;
-};
-
-constexpr bool operator<(const sym_entry& k1, const sym_entry& k2) {
-    return k1.hash < k2.hash;
-}
-
-static constexpr uint32_t elf_gnu_hash(const char* s) {
-    uint32_t h = 0x1505;
-    for(unsigned char c = *s; c != '\0'; c = *++s) h = (h << 5) + h + c;
-    return h;
-}
 
 /**
  * 
@@ -85,6 +67,8 @@ static const constexpr auto elf_api_table = sort(create_array_t<sym_entry>(
     API_METHOD(view_port_update, void, (ViewPort*)),
     API_METHOD(view_port_enabled_set, void, (ViewPort*, bool)),
     API_METHOD(view_port_free, void, (ViewPort*))));
+
+static_assert(!has_hash_collisions(elf_api_table), "Detected API method hash collision!");
 
 extern "C" bool elf_resolve_from_hashtable(const char* name, Elf32_Addr* address) {
     bool result = false;
