+++ applications/elf-loader/elf-cpp/elf-hashtable.h
@@ -6,6 +6,12 @@
 extern "C" {
 #endif
 
+/**
+ * Get function address by function name
+ * @param name function name
+ * @param address output for function address
+ * @return true if the table contains a function
+ */
 bool elf_resolve_from_hashtable(const char* name, Elf32_Addr* address);
 
 #ifdef __cplusplus
