+++ applications/elf-loader/elf-loader-app.c
@@ -2,7 +2,6 @@
 #include <storage/storage.h>
 #include <dialogs/dialogs.h>
 #include "elf-lib/elf-loader.h"
-#include "elf-loader-sys-api.h"
 #include "elf-cpp/elf-hashtable.h"
 
 #define APP_PATH "/ext/plugin.elf"
