--- applications/elf-loader/elf-loader-sys-api.h
@@ -1,3 +1,0 @@
-#pragma once
-
-typedef void(entry_t)(void);
